Extracts serializer pool and backend
